agent/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & ORM
sqlalchemy==2.0.23

# Authentication & Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1

# HuggingFace Models
transformers==4.35.2
torch==2.1.1
huggingface-hub==0.19.4

# MCP Framework
mcp==1.0.1
pydantic-core==2.14.1

# Web Search & Content Fetching
duckduckgo-search==3.9.11
requests==2.31.0
beautifulsoup4==4.12.2

# PowerPoint
python-pptx==0.6.23

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1
